Rollback to default admin and improve tabular inline
Material admin is going to be discontinued in Django 2, and I was not
super into it anyway. So bye bye!

The “raw_id_fields” option replaces the dropdown with search fields,
avoiding the cost of loading all enrollments / assignment tasks
multiple times from database (that was super slow). And as a plus, I
get a link to another admin page in each tabular inline row.